Modularize IoT telemetry helpers and docs
