+++ src/servers/iot/main.py
@@ -1,12 +1,7 @@
-import base64
-import binascii
-import json
 import logging
-import math
 import os
-from dataclasses import dataclass
-from datetime import datetime, timezone
-from typing import Any, Dict, Iterator, List, Optional, Tuple, Union
+from datetime import datetime
+from typing import Any, Dict, List, Optional, Union
 
 import couchdb3
 from dotenv import load_dotenv
@@ -22,13 +17,26 @@ from servers.iot.models import (
     FindAssetsResult,
     HistoryResult,
     LatestReadingResult,
-    SensorCoverage,
     SensorCoverageResult,
-    SensorStat,
     SensorStatsResult,
     SensorsResult,
     SitesResult,
     StreamExtentResult,
+)
+from servers.iot.telemetry_helper import (
+    _SensorAccumulator,
+    _SensorCoverageAccumulator,
+    _TimestampHandlingError,
+    _coerce_finite_number,
+    _decode_history_cursor,
+    _encode_history_cursor,
+    _history_cursor_context,
+    _history_observation,
+    _iter_records,
+    _iter_records_in_window,
+    _parse_iso_timestamp,
+    _timestamp_age_seconds,
+    _validate_dates,
 )
 
 load_dotenv()
@@ -73,16 +81,11 @@ except Exception as e:
 mcp = FastMCP(
     "iot",
     instructions=(
-        "IoT asset registry and telemetry record tools. Use sites() to discover site names, "
-        "asset_ids() for bare assetnum values at a site, asset_detail() for one asset's "
-        "registry details, assets() for registry metadata with optional assettype filtering, "
-        "find_assets_by_sensors() to find assets by installed or measured sensors, "
-        "installed_sensors() for registry sensor inventory, and measured_sensors() for "
-        "observed telemetry fields. Use stream_extent() to inspect telemetry time bounds "
-        "and record counts before planning larger telemetry reads. Use history() for paged "
-        "observations and latest_reading() for the newest values. Use sensor_coverage() for "
-        "per-field non-null counts and time coverage, and sensor_stats() for numeric summaries "
-        "without returning raw telemetry rows."
+        "Read-only tools for discovering IoT assets and analyzing telemetry. Resolve valid "
+        "sites, assets, and sensor names before querying telemetry. Prefer summary tools for "
+        "extent, coverage, and statistics; request history only when raw chronological "
+        "observations are needed. ISO 8601 windows are half-open and must use timezone "
+        "offsets consistently."
     ),
 )
 
@@ -95,123 +98,6 @@ _registry_sites_cache: Optional[List[str]] = None
 _sensor_list_cache: Dict[str, List[str]] = {}
 
 
-class _TimestampHandlingError(ValueError):
-    pass
-
-
-@dataclass
-class _SensorAccumulator:
-    count: int = 0
-    null_count: int = 0
-    min_value: Optional[float] = None
-    max_value: Optional[float] = None
-    mean_value: float = 0.0
-    squared_deviation_sum: float = 0.0
-    first_timestamp: Optional[str] = None
-    last_timestamp: Optional[str] = None
-    first_datetime: Optional[datetime] = None
-    last_datetime: Optional[datetime] = None
-
-    def add_invalid(self) -> None:
-        self.null_count += 1
-
-    def add_numeric(
-        self, value: float, timestamp: str, timestamp_dt: datetime
-    ) -> None:
-        self.count += 1
-        delta = value - self.mean_value
-        self.mean_value += delta / self.count
-        self.squared_deviation_sum += delta * (value - self.mean_value)
-        self.min_value = (
-            value if self.min_value is None else min(self.min_value, value)
-        )
-        self.max_value = (
-            value if self.max_value is None else max(self.max_value, value)
-        )
-        if self.first_datetime is None or timestamp_dt < self.first_datetime:
-            self.first_datetime = timestamp_dt
-            self.first_timestamp = timestamp
-        if self.last_datetime is None or timestamp_dt > self.last_datetime:
-            self.last_datetime = timestamp_dt
-            self.last_timestamp = timestamp
-
-    def result(self, sensor: str) -> SensorStat:
-        mean = None
-        stddev = None
-        if self.count:
-            if math.isfinite(self.mean_value):
-                mean = self.mean_value
-            variance = self.squared_deviation_sum / self.count
-            if math.isfinite(variance):
-                stddev = math.sqrt(max(variance, 0.0))
-        return SensorStat(
-            sensor=sensor,
-            count=self.count,
-            null_count=self.null_count,
-            min=self.min_value,
-            max=self.max_value,
-            mean=mean,
-            stddev=stddev,
-            first_timestamp=self.first_timestamp,
-            last_timestamp=self.last_timestamp,
-        )
-
-
-@dataclass
-class _SensorCoverageAccumulator:
-    non_null_count: int = 0
-    first_timestamp: Optional[str] = None
-    last_timestamp: Optional[str] = None
-    first_datetime: Optional[datetime] = None
-    last_datetime: Optional[datetime] = None
-
-    def add_non_null(self, timestamp: str, timestamp_dt: datetime) -> None:
-        self.non_null_count += 1
-        if self.first_datetime is None or timestamp_dt < self.first_datetime:
-            self.first_datetime = timestamp_dt
-            self.first_timestamp = timestamp
-        if self.last_datetime is None or timestamp_dt > self.last_datetime:
-            self.last_datetime = timestamp_dt
-            self.last_timestamp = timestamp
-
-    def result(self, sensor: str) -> SensorCoverage:
-        return SensorCoverage(
-            sensor=sensor,
-            non_null_count=self.non_null_count,
-            first_timestamp=self.first_timestamp,
-            last_timestamp=self.last_timestamp,
-        )
-
-
-def _iter_records(
-    selector: Dict[str, Any],
-    fields: Optional[List[str]] = None,
-    sort: Optional[List[Dict[str, str]]] = None,
-    page_size: int = PAGE_SIZE,
-) -> Iterator[Dict[str, Any]]:
-    """Yield telemetry records matching a selector across paged database results."""
-    if not iot_db:
-        return
-    if sort is None:
-        sort = [{"asset_id": "asc"}, {"timestamp": "asc"}]
-    bookmark: Optional[str] = None
-    while True:
-        kwargs: Dict[str, Any] = {"limit": page_size, "sort": sort}
-        if fields is not None:
-            kwargs["fields"] = fields
-        if bookmark is not None:
-            kwargs["bookmark"] = bookmark
-        res = iot_db.find(selector, **kwargs)
-        docs = res.get("docs", [])
-        if not docs:
-            break
-        for doc in docs:
-            yield doc
-        bookmark = res.get("bookmark")
-        if bookmark is None or len(docs) < page_size:
-            break
-
-
 def get_sensor_list(asset_id: str) -> List[str]:
     """Return sorted telemetry field names observed across all records for an asset."""
     if asset_id in _sensor_list_cache:
@@ -221,7 +107,7 @@ def get_sensor_list(asset_id: str) -> List[str]:
     try:
         found = set()
         seen = False
-        for doc in _iter_records({"asset_id": asset_id}):
+        for doc in _iter_records(iot_db, {"asset_id": asset_id}):
             seen = True
             found.update(key for key in doc.keys() if key not in RESERVED_FIELDS)
         if not seen:
@@ -297,191 +183,29 @@ def _installed_sensors(asset_id: str, site_name: Optional[str] = None) -> List[s
         return []
 
 
-def _parse_iso_timestamp(value: Any) -> Optional[datetime]:
-    """Parse one ISO 8601 timestamp, returning None for invalid values."""
-    if not isinstance(value, str):
-        return None
-    try:
-        return datetime.fromisoformat(value)
-    except (TypeError, ValueError):
-        return None
-
-
-def _is_timezone_aware(value: datetime) -> bool:
-    return value.utcoffset() is not None
-
-
-def _validate_dates(start: Optional[str], end: Optional[str]) -> Optional[str]:
-    """Return None when the optional ISO 8601 bounds are valid."""
-    start_dt = _parse_iso_timestamp(start) if start is not None else None
-    end_dt = _parse_iso_timestamp(end) if end is not None else None
-    if start is not None and start_dt is None:
-        return "Invalid date format for start (expected ISO 8601)"
-    if end is not None and end_dt is None:
-        return "Invalid date format for end (expected ISO 8601)"
-    if start_dt is not None and end_dt is not None:
-        if _is_timezone_aware(start_dt) != _is_timezone_aware(end_dt):
-            return "start and end must use matching timezone awareness"
-        if start_dt >= end_dt:
-            return "start >= end"
-    return None
-
-
-def _iter_records_in_window(
-    selector: Dict[str, Any],
-    start_dt: Optional[datetime],
-    end_dt: Optional[datetime],
-    fields: Optional[List[str]] = None,
-) -> Iterator[Tuple[Dict[str, Any], str, datetime]]:
-    """Yield timestamped records in a parsed half-open time window."""
-    stream_is_aware: Optional[bool] = None
-    for doc in _iter_records(selector, fields=fields):
-        timestamp = doc.get("timestamp")
-        if timestamp is None:
-            continue
-        timestamp_dt = _parse_iso_timestamp(timestamp)
-        if timestamp_dt is None:
-            raise _TimestampHandlingError(
-                "telemetry record has an invalid ISO 8601 timestamp"
-            )
-
-        timestamp_is_aware = _is_timezone_aware(timestamp_dt)
-        if stream_is_aware is None:
-            stream_is_aware = timestamp_is_aware
-            for bound in (start_dt, end_dt):
-                if (
-                    bound is not None
-                    and _is_timezone_aware(bound) != stream_is_aware
-                ):
-                    raise _TimestampHandlingError(
-                        "timestamp bounds must use the same timezone awareness "
-                        "as telemetry timestamps"
-                    )
-        elif timestamp_is_aware != stream_is_aware:
-            raise _TimestampHandlingError(
-                "telemetry timestamps use mixed timezone awareness"
-            )
-
-        if start_dt is not None and timestamp_dt < start_dt:
-            continue
-        if end_dt is not None and timestamp_dt >= end_dt:
-            continue
-        yield doc, timestamp, timestamp_dt
-
-
-def _coerce_finite_number(value: Any) -> Optional[float]:
-    """Return a finite float for numeric values and numeric strings."""
-    if isinstance(value, bool):
-        return None
-    try:
-        number = float(value)
-    except (TypeError, ValueError):
-        return None
-    return number if math.isfinite(number) else None
-
-
-def _history_cursor_context(
-    site_name: str,
-    asset_id: str,
-    start: Optional[str],
-    end: Optional[str],
-    sensors: Optional[List[str]],
-) -> Dict[str, Any]:
-    return {
-        "site_name": site_name,
-        "asset_id": asset_id,
-        "start": start,
-        "end": end,
-        "sensors": sensors,
-    }
-
-
-def _encode_history_cursor(offset: int, context: Dict[str, Any]) -> str:
-    payload = json.dumps(
-        {"version": 1, "offset": offset, "context": context},
-        separators=(",", ":"),
-        sort_keys=True,
-    ).encode("utf-8")
-    return base64.urlsafe_b64encode(payload).decode("ascii").rstrip("=")
-
-
-def _decode_history_cursor(
-    cursor: str, expected_context: Dict[str, Any]
-) -> Tuple[Optional[int], Optional[str]]:
-    try:
-        padding = "=" * (-len(cursor) % 4)
-        payload = json.loads(
-            base64.urlsafe_b64decode(cursor + padding).decode("utf-8")
-        )
-    except (binascii.Error, UnicodeDecodeError, ValueError, json.JSONDecodeError):
-        return None, "invalid cursor"
-    if not isinstance(payload, dict) or payload.get("version") != 1:
-        return None, "invalid cursor"
-    offset = payload.get("offset")
-    if isinstance(offset, bool) or not isinstance(offset, int) or offset < 0:
-        return None, "invalid cursor"
-    if payload.get("context") != expected_context:
-        return None, "cursor does not match history query"
-    return offset, None
-
-
-def _history_observation(
-    doc: Dict[str, Any],
-    asset_id: str,
-    timestamp: str,
-    sensors: Optional[List[str]],
-) -> Dict[str, Any]:
-    observation: Dict[str, Any] = {
-        "asset_id": asset_id,
-        "timestamp": timestamp,
-    }
-    if sensors is None:
-        observation.update(
-            {
-                field: value
-                for field, value in doc.items()
-                if field not in RESERVED_FIELDS
-            }
-        )
-    else:
-        observation.update(
-            {field: doc[field] for field in sensors if field in doc}
-        )
-    return observation
-
-
-def _timestamp_age_seconds(timestamp_dt: datetime) -> float:
-    if not _is_timezone_aware(timestamp_dt):
-        timestamp_dt = timestamp_dt.replace(tzinfo=timezone.utc)
-    return (
-        datetime.now(timezone.utc) - timestamp_dt.astimezone(timezone.utc)
-    ).total_seconds()
-
-
 @mcp.tool(title="List Sites")
 def sites() -> SitesResult:
-    """List known site names from the asset registry.
+    """List sorted site identifiers available in the asset registry.
 
     Returns:
-        SitesResult: `sites` contains sorted distinct `siteid` values from
-        asset profiles. If the registry database is unavailable or has no sites,
-        `sites` falls back to `["MAIN"]`.
+        SitesResult: Distinct site identifiers, or `["MAIN"]` when none are
+        available.
     """
     return SitesResult(sites=known_sites())
 
 
 @mcp.tool(title="List Asset IDs")
 def asset_ids(site_name: str) -> Union[AssetsResult, ErrorResult]:
-    """List only asset identifiers for one site.
+    """List asset identifiers registered at one site.
+
+    Use `assets()` when registry metadata is also needed.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
             discover valid site ids.
 
     Returns:
-        AssetsResult: Contains `site_name`, `total_assets`, `assets`, and
-        `message`. The `assets` field is a sorted list of asset registry
-        `assetnum` values.
+        AssetsResult: Sorted asset identifiers and their count.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -509,9 +233,8 @@ def asset_ids(site_name: str) -> Union[AssetsResult, ErrorResult]:
 def asset_detail(site_name: str, asset_id: str) -> Union[AssetDetail, ErrorResult]:
     """Return registry details for one asset.
 
-    This tool reads the asset registry from `asset_db` (`ASSET_DBNAME`, default
-    `asset`) and returns nameplate fields for one asset record. Use
-    `installed_sensors()` when you only need the registry sensor inventory.
+    Includes identity, description, type, status, location, installation date,
+    vintage, and installed-sensor count. Use `installed_sensors()` for names.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
@@ -519,9 +242,8 @@ def asset_detail(site_name: str, asset_id: str) -> Union[AssetDetail, ErrorResul
         asset_id: Exact asset id from `asset_ids()`, such as `Chiller 6`.
 
     Returns:
-        AssetDetail: Contains `site_name`, `asset_id`, `description`,
-        `assettype`, `status`, `location`, `installdate`, `vintage`,
-        `n_installed_sensors`, and `message`.
+        AssetDetail: The available registry fields; optional values are null
+        when absent.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -582,12 +304,10 @@ def asset_detail(site_name: str, asset_id: str) -> Union[AssetDetail, ErrorResul
 def measured_sensors(
     site_name: str, asset_id: str
 ) -> Union[SensorsResult, ErrorResult]:
-    """List telemetry fields actually measured for one asset.
+    """List measurement fields observed in an asset's telemetry stream.
 
-    This tool reads IoT records from `iot_db` (`IOT_DBNAME`, default `iot`).
-    It scans records matching `asset_id` and returns the union of observed
-    measurement fields. Use `installed_sensors()` instead when you need the
-    asset registry inventory.
+    Returns the sorted union across the full stream, excluding record metadata.
+    Use `installed_sensors()` for assigned sensors, which may differ.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
@@ -595,10 +315,7 @@ def measured_sensors(
         asset_id: Exact asset id from `asset_ids()`, such as `Chiller 6`.
 
     Returns:
-        SensorsResult: Contains `site_name`, `asset_id`, `total_sensors`,
-        `sensors`, and `message`. The `sensors` field is a sorted list of
-        observed telemetry fields, excluding record metadata such as `_id`,
-        `_rev`, `asset_id`, `timestamp`, `dataset`, `type`, and `doctype`.
+        SensorsResult: Observed measurement field names and their count.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -625,12 +342,10 @@ def measured_sensors(
 def installed_sensors(
     site_name: str, asset_id: str
 ) -> Union[SensorsResult, ErrorResult]:
-    """List sensor names installed on one asset according to the registry.
+    """List sensor names assigned to an asset in the registry.
 
-    This tool reads the asset registry from `asset_db` (`ASSET_DBNAME`, default
-    `asset`) and returns the asset record's `sensors` inventory. Use
-    `measured_sensors()` instead when you need fields actually observed in IoT
-    telemetry records.
+    Use `measured_sensors()` for fields observed in telemetry; the two lists may
+    differ.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
@@ -638,9 +353,7 @@ def installed_sensors(
         asset_id: Exact asset id from `asset_ids()`, such as `Chiller 6`.
 
     Returns:
-        SensorsResult: Contains `site_name`, `asset_id`, `total_sensors`,
-        `sensors`, and `message`. The `sensors` field preserves the registry
-        sensor inventory order from the asset record.
+        SensorsResult: Sensor names in registry order and their count.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -673,19 +386,20 @@ def installed_sensors(
 def assets(
     site_name: str, assettype: Optional[str] = None
 ) -> Union[AssetsWithMetadataResult, ErrorResult]:
-    """List asset registry records for one site with compact metadata.
+    """List assets at one site with compact registry metadata.
+
+    Use `asset_ids()` for identifiers only and `asset_detail()` for more fields
+    about one asset.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
             discover valid site ids.
-        assettype: Optional exact asset type filter, such as `PUMP` or
-            `COMPRESSOR`.
+        assettype: Optional exact, case-sensitive asset type filter, such as
+            `PUMP` or `COMPRESSOR`.
 
     Returns:
-        AssetsWithMetadataResult: Contains `site_name`, `total_assets`,
-        `assets`, and `message`. Each item in `assets` includes `asset_id`
-        (the registry `assetnum`), `description`, `assettype`, `vintage`, and
-        `n_sensors`.
+        AssetsWithMetadataResult: Assets sorted by identifier, with description,
+        type, vintage, and installed-sensor count.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -734,24 +448,24 @@ def find_assets_by_sensors(
     substring: bool = False,
     source: str = "measured",
 ) -> Union[FindAssetsResult, ErrorResult]:
-    """Return assets at a site that match the requested sensor names.
+    """Find site assets by installed or measured sensor names.
 
-    The search is limited to assets registered at `site_name`. Duplicate query
-    sensors are ignored while preserving the first occurrence order.
+    Searches only assets registered at `site_name`. Duplicate queries are
+    removed while preserving first occurrence order.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
             discover valid site ids.
-        sensors: One or more sensor names or, when `substring` is true, sensor
-            name fragments to search for.
-        match: `all` requires every listed sensor; `any` requires at least one.
-        substring: If true, match sensor names case-insensitively by substring.
-        source: `measured` checks telemetry fields; `installed` checks the
-            asset registry inventory.
+        sensors: Sensor names, or fragments when `substring` is true.
+        match: `all` requires every query; `any` requires at least one.
+        substring: Use case-insensitive fragment matching instead of exact,
+            case-sensitive matching.
+        source: `measured` searches observed fields; `installed` searches
+            registry assignments. Defaults to `measured`.
 
     Returns:
-        FindAssetsResult: Contains the deduplicated query sensors, matching
-        asset ids, and the concrete sensor names that matched each asset.
+        FindAssetsResult: Query settings and matching assets with the concrete
+        sensor names found.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -832,11 +546,11 @@ def stream_extent(
 ) -> Union[StreamExtentResult, ErrorResult]:
     """Inspect the size and timestamp span of an asset's telemetry stream.
 
-    Optionally restrict the stream to one measured sensor and/or an ISO 8601
-    window. Sensor-filtered records count only when the field is present and
-    non-null. The window is half-open: `start` is inclusive and `end` is
-    exclusive. Bounds and telemetry timestamps must consistently include or
-    omit timezone offsets; explicit offsets may differ and compare by instant.
+    Optionally filter by one measured sensor and a half-open ISO 8601 window:
+    `start` is inclusive and `end` is exclusive. Sensor-filtered records count
+    only when the field is present and non-null. Bounds and record timestamps
+    must consistently include or omit timezone offsets; aware timestamps are
+    compared by instant even when their offsets differ.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
@@ -849,11 +563,9 @@ def stream_extent(
         end: Optional exclusive ISO 8601 date or datetime upper bound.
 
     Returns:
-        StreamExtentResult: `start_time` and `end_time` are the earliest and
-        latest matching timestamps. `total_records` is the matching count,
-        `exceeds_page_limit` indicates more than one server page, and
-        `approx_interval_seconds` is the average spacing implied by the span and
-        count.
+        StreamExtentResult: Earliest and latest matching timestamps, record
+        count, whether the count exceeds one 1000-record page, and average
+        interval implied by the span and count.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -885,7 +597,7 @@ def stream_extent(
     total_records = 0
     try:
         for _, timestamp, timestamp_dt in _iter_records_in_window(
-            selector, start_dt, end_dt, fields=["timestamp"]
+            iot_db, selector, start_dt, end_dt, fields=["timestamp"]
         ):
             if first_datetime is None or timestamp_dt < first_datetime:
                 first_datetime = timestamp_dt
@@ -943,33 +655,29 @@ def history(
 ) -> Union[HistoryResult, ErrorResult]:
     """Return one chronological page of telemetry observations for an asset.
 
-    Optional bounds form a half-open ISO 8601 window: `start` is inclusive and
-    `end` is exclusive. Use `sensors` to project exact fields from
-    `measured_sensors()`. Every row includes `asset_id` and `timestamp`; other
-    metadata and internal record identifiers are excluded. Omit `sensors` to
-    return every measured field present in each row.
+    Optional bounds form a half-open ISO 8601 window. Use `sensors` to project
+    exact measured fields; omit it to return all measured fields present in each
+    row. Every row includes `asset_id` and `timestamp` and excludes other record
+    metadata.
 
-    `limit` must be between 1 and 1000. Leave `cursor` unset for the first page.
-    When `has_more` is true, repeat the same query arguments with `next_cursor`.
-    Cursors are opaque and bound to the site, asset, window, and sensor list.
-    Timestamp offset rules match `stream_extent()`; history returns an error if
-    timestamp representations cannot be emitted in chronological order.
+    Leave `cursor` unset for the first page. When `has_more` is true, repeat the
+    same query with `next_cursor`. Cursors are opaque and query-bound. Timestamp
+    offset rules match `stream_extent()`.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
             discover valid site ids.
         asset_id: Exact asset id from `asset_ids()`, such as `Chiller 6`.
-        start: Optional inclusive ISO 8601 date or datetime lower bound.
-        end: Optional exclusive ISO 8601 date or datetime upper bound.
+        start: Optional inclusive ISO 8601 lower bound.
+        end: Optional exclusive ISO 8601 upper bound.
         sensors: Optional non-empty list of exact measured field names. Duplicate
             names are removed while preserving order.
-        limit: Maximum observations in this page, from 1 through 1000.
+        limit: Maximum observations in this page, from 1 to 1000.
         cursor: Opaque `next_cursor` from the previous page of the same query.
 
     Returns:
-        HistoryResult: `observations` contains the projected chronological rows,
-        `returned` is the page size, and `has_more`/`next_cursor` control paging.
-        `start` and `end` echo the requested window.
+        HistoryResult: Projected observations, returned count, echoed bounds,
+        and cursor fields for the next page.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -1030,7 +738,7 @@ def history(
     previous_datetime: Optional[datetime] = None
     try:
         for doc, timestamp, timestamp_dt in _iter_records_in_window(
-            selector, start_dt, end_dt, fields=fields
+            iot_db, selector, start_dt, end_dt, fields=fields
         ):
             if previous_datetime is not None and timestamp_dt < previous_datetime:
                 raise _TimestampHandlingError(
@@ -1045,7 +753,11 @@ def history(
                 break
             observations.append(
                 _history_observation(
-                    doc, asset_id, timestamp, selected_sensors
+                    doc,
+                    asset_id,
+                    timestamp,
+                    selected_sensors,
+                    RESERVED_FIELDS,
                 )
             )
             matched_records += 1
@@ -1084,11 +796,9 @@ def latest_reading(
 ) -> Union[LatestReadingResult, ErrorResult]:
     """Return the newest telemetry observation for an asset.
 
-    Omit `sensor` to return every measured field present in the newest record.
-    Provide an exact field name from `measured_sensors()` to return the newest
-    record where that field is present and non-null. The newest record is chosen
-    by parsed timestamp rather than string order, using the same timestamp rules
-    as `stream_extent()`.
+    Omit `sensor` for all measured fields in the newest record. When a sensor is
+    provided, returns the newest record where that field is present and non-null.
+    Records are compared by parsed timestamp using the `stream_extent()` rules.
 
     `age_seconds` is the difference between the current UTC time and the reading
     timestamp. Offset-free timestamps are interpreted as UTC; a negative value
@@ -1102,8 +812,7 @@ def latest_reading(
             metadata fields are not valid sensor names.
 
     Returns:
-        LatestReadingResult: Contains the selected observation timestamp,
-        requested sensor values, age in seconds, and a compact message.
+        LatestReadingResult: Selected timestamp, values, and age in seconds.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -1139,7 +848,7 @@ def latest_reading(
     latest_datetime: Optional[datetime] = None
     try:
         for doc, timestamp, timestamp_dt in _iter_records_in_window(
-            selector, None, None, fields=fields
+            iot_db, selector, None, None, fields=fields
         ):
             if latest_datetime is None or timestamp_dt > latest_datetime:
                 latest_doc = doc
@@ -1182,12 +891,9 @@ def sensor_coverage(
 ) -> Union[SensorCoverageResult, ErrorResult]:
     """Summarize non-null observation coverage for every measured sensor.
 
-    Scans the asset's full timestamped stream; use `stream_extent()` first when
-    you need to estimate its size. `non_null_count` includes present non-null
-    values of any type. Fields observed only with null values are returned with
-    zero count and null time bounds. First and last timestamps are selected
-    chronologically. Mixed offset-aware and offset-free timestamps return an
-    error.
+    Scans the full timestamped stream. `non_null_count` includes present,
+    non-null values of any type. Fields observed only with null values have zero
+    count and null bounds. Timestamp offset rules match `stream_extent()`.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
@@ -1195,9 +901,8 @@ def sensor_coverage(
         asset_id: Exact asset id from `asset_ids()`, such as `Chiller 6`.
 
     Returns:
-        SensorCoverageResult: `docs_scanned` is the number of timestamped records
-        examined. `sensors` is sorted by field name and contains each field's
-        non-null count and earliest/latest non-null timestamps.
+        SensorCoverageResult: Timestamped records scanned and per-sensor counts
+        and chronological bounds, sorted by field name.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -1212,7 +917,7 @@ def sensor_coverage(
     docs_scanned = 0
     try:
         for doc, timestamp, timestamp_dt in _iter_records_in_window(
-            selector, None, None
+            iot_db, selector, None, None
         ):
             docs_scanned += 1
             for field, value in doc.items():
@@ -1256,15 +961,13 @@ def sensor_stats(
 ) -> Union[SensorStatsResult, ErrorResult]:
     """Compute numeric statistics for one or all measured sensors.
 
-    Omit `sensor` to summarize every field returned by `measured_sensors()`.
-    Optional bounds form a half-open ISO 8601 window: `start` is inclusive and
-    `end` is exclusive. Timestamp offset rules match `stream_extent()`.
+    Omit `sensor` to summarize every measured field. Optional bounds form a
+    half-open ISO 8601 window. Timestamp offset rules match `stream_extent()`.
 
     Finite numbers and numeric strings contribute to the numeric statistics.
     Present null, boolean, non-numeric, and non-finite values contribute to
     `null_count`; missing fields contribute to neither count. `stddev` is the
     population standard deviation: `0.0` for one numeric value and null for none.
-    First and last timestamps identify the earliest and latest numeric samples.
 
     Args:
         site_name: Exact site id to query, such as `MAIN`. Use `sites()` to
@@ -1272,14 +975,12 @@ def sensor_stats(
         asset_id: Exact asset id from `asset_ids()`, such as `Chiller 6`.
         sensor: Optional exact field name from `measured_sensors()`. Omit it to
             summarize all measured fields.
-        start: Optional inclusive ISO 8601 date or datetime lower bound.
-        end: Optional exclusive ISO 8601 date or datetime upper bound.
+        start: Optional inclusive ISO 8601 lower bound.
+        end: Optional exclusive ISO 8601 upper bound.
 
     Returns:
-        SensorStatsResult: `stats` contains one entry per requested field, sorted
-        by field name when `sensor` is omitted. Each entry includes numeric and
-        null counts, range, mean, population standard deviation, and
-        earliest/latest numeric sample timestamps.
+        SensorStatsResult: Per-field numeric and null counts, range, mean,
+        population standard deviation, and earliest/latest numeric timestamps.
     """
     if not _is_known_site(site_name):
         return ErrorResult(error=f"unknown site {site_name}")
@@ -1315,6 +1016,7 @@ def sensor_stats(
     records_in_window = 0
     try:
         for doc, timestamp, timestamp_dt in _iter_records_in_window(
+            iot_db,
             selector,
             start_dt,
             end_dt,
